[TM-3423] improve code, and move bbox to adapters in mapbox
